# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ONG API by Boniich",
        "description": "I make this API to youtube",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/slides": {
            "get": {
                "tags": [
                    "Slides"
                ],
                "summary": "Display a list of slides",
                "operationId": "10978508e846810c82bcfc1e069a1399",
                "responses": {
                    "200": {
                        "description": "Slides retrived successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "format": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "name": "Slide",
                                                    "description": "Description of slide",
                                                    "image": "slides/12345593.png",
                                                    "user_id": 1,
                                                    "users": {
                                                        "id": 1,
                                                        "name": "Boniich",
                                                        "email": "boniich@gmail.com",
                                                        "address": "901 Collier Forks",
                                                        "profile_image": null
                                                    }
                                                }
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Slides retrived successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Slides"
                ],
                "summary": "Create a new slide",
                "operationId": "28f7918fc4969cafe985c423f3a8383e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name of slide",
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "description": {
                                        "description": "description of slide",
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "image": {
                                        "description": "image of slide",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Slide created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "format": "boolean",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": {
                                                "id": 1,
                                                "name": "Slide",
                                                "description": "Description of slide",
                                                "image": "slides/12345593.png",
                                                "user_id": 1
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Slide created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "format": "boolean",
                                            "example": "false"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "format": "integer",
                                            "example": "401"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "format": "boolean",
                                            "example": "false"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}